+++ code/classeine-lib/clsn/ui/amatista/AmatistaUISkin.cpp
@@ -14,6 +14,7 @@
 
 #include "clsn/ui/renderers/ButtonRenderer.h"
 #include "clsn/ui/renderers/CheckBoxRenderer.h"
+#include "clsn/ui/renderers/DualContainerRenderer.h"
 #include "clsn/ui/renderers/HBoxContainerRenderer.h"
 #include "clsn/ui/renderers/LabelRenderer.h"
 #include "clsn/ui/renderers/RadioButtonRenderer.h"
@@ -29,6 +30,7 @@ namespace
     {
         skin.makeRendererAndAdd<Button, ButtonRenderer>();
         skin.makeRendererAndAdd<CheckBox, CheckBoxRenderer>();
+        skin.makeRendererAndAdd<DualContainer, DualContainerRenderer>();
         skin.makeRendererAndAdd<HBoxContainer, HBoxContainerRenderer>();
         skin.makeRendererAndAdd<Label, LabelRenderer>();
         skin.makeRendererAndAdd<RadioButton, RadioButtonRenderer>();